fixed ordering +pagination, removed ME4, inprog orderByPopularity
Fixed ordering by Name, Author, Date, etc. these all now work off of replica indices in Algolia.

Pagination also now has a variable so that it is "optional" as long as it is used correctly with the relevant controller logic. NOTE: do some user testing as to whether or not the pagination should be used.

Removed the option of ME4 from databases, seeders, csvs. Also make the seeder creation more fluid, with the "test" admin user being added to staff, although he/she must still register.

In progress of adding "orderByPopularity". This involves creating a new database column for projects so that aloglia can access it. For now it shall simply just mimic the number of likes.