Fixed bug with search engine not showing created projects
Algolia is SUPER finnicky with fields, (i.e. null =/= false =/= 0). AND the fields must be assigned on creation -> i.e. cannot be left to default assignment.

Also there is a wierd interaction in which laravel deletes the object and algolia search doesnt show it BUT it keeps the index on the search engine??? Wierd but doesn't cause any visible problems to users - only slightly ineffiecient indexing.